all tests ammended
