--- app/models/pub_challenge.rb
@@ -1,8 +1,0 @@
-class PubChallenge < ActiveRecord::Base
-  belongs_to :challenge
-  belongs_to :pub
-
-  def self.get_challenges
-    Challenge.all.shuffle[0..8]
-  end
-end
