+++ spec/controllers/crawls_controller_spec.rb
@@ -19,7 +19,9 @@ RSpec.describe CrawlsController, type: :controller do
     end
 
     it 'gets a new of challenge' do
-      expect(PubChallenge).to receive(:get_challenges).and_return( [challenge] )
+      Challenge.create
+
+      expect(CrawlChallenge).to receive(:get_challenges).and_return( [challenge] )
       post :create, crawl: { "name": crawl_name, "start_postcode": postcode}
     end
   end
