+++ spec/models/challenge_spec.rb
@@ -1,5 +1,5 @@
 require 'rails_helper'
 
 RSpec.describe Challenge, type: :model do
-  it{ is_expected.to have_many(:pub_challenges) }
+  it{ is_expected.to have_many(:crawl_challenges) }
 end
