+++ spec/models/pub_spec.rb
@@ -1,5 +1,5 @@
 require 'rails_helper'
 
 RSpec.describe Pub, type: :model do
-  it{ is_expected.to have_many(:pub_challenges) }
+  it{ is_expected.not_to have_many(:pub_challenges) }
 end
